--- src/Customer.java
@@ -1,24 +1,0 @@
-public interface Customer {
-
-    public String getCode();
-    public void setCode(String code);
-
-    public String getName();
-
-    public String getIdNumber();
-
-    public String getPostalAddress();
-    public void setPostalAddress(String postalAddress);
-
-    public String getCity();
-    public void setCity(String city);
-
-    public String getProvince();
-    public void setProvince(String province);
-
-    public String getEmail();
-    public void setEmail(String email);
-
-    public String getPhoneNumber();
-    public void setPhoneNumber(String phoneNumber);
-}
