--- src/SalesLine.java
@@ -1,13 +1,0 @@
-public interface SalesLine {
-
-    //Metodos
-    public String getCode();
-    public void setCode(String code);
-
-    public String getName();
-    public void setName(String name);
-
-    public Double getPrice();
-    public void setPrice(Double price);
-
-}
