--- src/VatRate.java
@@ -1,9 +1,0 @@
-public class VatRate {
-    //Atributos
-    private Integer type;
-
-    //Metodos
-    public Integer getType(){return type;}
-    public void setType(Integer type){this.type = type;}
-
-}
