+++ src/com/iesam/chispas/data/InvoiceDataStore.java
@@ -1,0 +1,7 @@
+package com.iesam.chispas.data;
+import com.iesam.chispas.domain.models.VatRate;
+import java.util.List;
+
+public interface InvoiceDataStore {
+
+}
