+++ src/com/iesam/chispas/domain/models/Customer.java
@@ -1,0 +1,26 @@
+package com.iesam.chispas.domain.models;
+
+public interface Customer {
+
+    public String getCode();
+    public void setCode(String code);
+
+    public String getName();
+
+    public String getIdNumber();
+
+    public String getPostalAddress();
+    public void setPostalAddress(String postalAddress);
+
+    public String getCity();
+    public void setCity(String city);
+
+    public String getProvince();
+    public void setProvince(String province);
+
+    public String getEmail();
+    public void setEmail(String email);
+
+    public String getPhoneNumber();
+    public void setPhoneNumber(String phoneNumber);
+}
