+++ src/com/iesam/chispas/domain/models/SalesLine.java
@@ -1,0 +1,15 @@
+package com.iesam.chispas.domain.models;
+
+public interface SalesLine {
+
+    //Metodos
+    public String getCode();
+    public void setCode(String code);
+
+    public String getName();
+    public void setName(String name);
+
+    public Double getPrice();
+    public void setPrice(Double price);
+
+}
