+++ src/com/iesam/chispas/domain/models/VatRate.java
@@ -1,0 +1,11 @@
+package com.iesam.chispas.domain.models;
+
+public class VatRate {
+    //Atributos
+    private Integer type;
+
+    //Metodos
+    public Integer getType(){return type;}
+    public void setType(Integer type){this.type = type;}
+
+}
