+++ src/com/iesam/chispas/domain/usecase/DeleteSalesLineUseCase.java
@@ -1,0 +1,12 @@
+package com.iesam.chispas.domain.usecase;
+
+import com.iesam.chispas.data.MenSalesLineDataStore;
+import com.iesam.chispas.data.SalesLineDataStore;
+import com.iesam.chispas.domain.models.SalesLine;
+
+public class DeleteSalesLineUseCase {
+    public void execute(SalesLine salesLine){
+        SalesLineDataStore salesLineDataStore = new MenSalesLineDataStore();
+        salesLineDataStore.deleteSalesLine(salesLine);
+    }
+}
