--- src/data/CustomerDataStore.java
@@ -1,10 +1,0 @@
-package data;
-
-public interface CustomerDataStore {
-    void saveCustomer(Customer customer);
-    void deleteCustomer(Customer customer);
-    void updateCustomer (Customer customer);
-    List<Customer> getAllCustomer();
-    Customer findById(String code);
-
-}
