--- src/data/InvoiceDataStore.java
@@ -1,8 +1,0 @@
-package data;
-
-public interface InvoiceDataStore {
-
-    List<VatRate> getAllVatRate();
-    VatRate getVatRateByType(Integer type);
-
-}
