--- src/data/MenInvoiceDataStore.java
@@ -1,16 +1,0 @@
-package data;
-
-public class MenInvoiceDataStore implements InvoiceDataStore{
-
-    private Map<String, Invoice> dataStore = new TreepMap<String, Invoice>();
-
-    @Override
-    public List<VatRate> getAllVatRate() {
-        System.out.prinln(dataStore.values().stream().toList());
-    }
-
-    @Override
-    public VatRate getVatRateByType(Integer type) {
-        return VatRate;
-    }
-}
