--- src/data/SalesLineDataStore.java
@@ -1,9 +1,0 @@
-package data;
-
-public interface SalesLineDataStore {
-    void saveSalesLine(SalesLine salesLine);
-    void deleteSalesLine(SalesLine salesLine);
-    void updateSalesLine(SalesLine salesLine);
-    List<SalesLine> getAllSalesLine();
-    Customer findById(String codeSalesLine);
-}
