--- src/domain/AddCustomerUseCase.java
@@ -1,10 +1,0 @@
-package domain;
-
-import data.CustomerDataStore;
-
-public class AddCustomerUseCase {
-    public void execute(Customer customer){
-        CustomerDataStore customerDataStore = new CustomerDataStore();
-        customerDataStore.saveCustomer(customer);
-    }
-}
