--- src/domain/AddSalesLineUseCase.java
@@ -1,11 +1,0 @@
-package domain;
-
-import data.SalesLineDataStore;
-
-public class AddSalesLineUseCase {
-    public void execute (SalesLine salesLine){
-        SalesLineDataStore salesLineDataStore = new SalesLineDataStore();
-        salesLineDataStore.saveSalesLine(salesLine);
-    }
-}
-
