--- src/domain/DeleteCustomerUseCase.java
@@ -1,10 +1,0 @@
-package domain;
-
-import data.CustomerDataStore;
-
-public class DeleteCustomerUseCase {
-    public void execute(Customer customer){
-        CustomerDataStore customerDataStore = new CustomerDataStore();
-        customerDataStore.deleteCustomer(customer);
-    }
-}
