--- src/domain/GetCustomerUseCase.java
@@ -1,10 +1,0 @@
-package domain;
-
-import data.CustomerDataStore;
-
-public class GetCustomerUseCase {
-    public Customer execute(String cod){
-        CustomerDataStore customerDataStore = new CustomerDataStore();
-        return customerDataStore.findById(cod);
-    }
-}
