--- src/domain/GetSalesLineUseCase.java
@@ -1,10 +1,0 @@
-package domain;
-
-import data.SalesLineDataStore;
-
-public class GetSalesLineUseCase {
-    public SalesLine execute(String code){
-        SalesLineDataStore salesLineDataStore = new SalesLineDataStore();
-        return salesLineDataStore.findById(code);
-    }
-}
