--- src/domain/ListCustomersUseCase.java
@@ -1,10 +1,0 @@
-package domain;
-
-import data.CustomerDataStore;
-
-public class ListCustomersUseCase {
-    public List<Customer> execute(){
-        CustomerDataStore customerDataStore = new CustomerDataStore();
-        return customerDataStore.getAllCustomer();
-    }
-}
