--- src/domain/ListSalesLineUseCase.java
@@ -1,10 +1,0 @@
-package domain;
-
-import data.SalesLineDataStore;
-
-public class ListSalesLineUseCase {
-    public List<SalesLine> execute(){
-        SalesLineDataStore salesLineDataStore = new SalesLineDataStore();
-        return salesLineDataStore.getAllSalesLine();
-    }
-}
