--- src/domain/UpdateCustomerUseCase.java
@@ -1,7 +1,0 @@
-package domain;
-
-public class UpdateCustomerUseCase {
-    public void execute(Customer customer){
-
-    }
-}
